Renamed the node_id member to local_node_id
We now distinguish the local node id (the one being used by the python environment) from the actual node id. This allows setting up the nodes using an identifier which is not related to the actual node id, as we cannot determine the id that will be generated on the KV, yet still have the actual node id once it is up and running.