In pbft keep tracks of the last pre-prepare that performed a governance request (#994)
This change creates a "chain" of pre-prepares that contain requests that performed some kind of governance operation. What this means that when a user gets their receipt (which covers a pre-prepare) they can make a determination if they want to audit the log and if they do they do not need to look into every pre-prepare.