V8 pipeline publishing artifact
Creating a script to build V8 and a pipeline config to run that,
publishing an artifact with the monolith version (ready for embedding).

The pipeline has two build options:
 * VERSION (mandatory): ex. 9.4.146.17
 * PUSH_ARTIFACT: true|false

We only need to publish the artifact once, but we can continue to build
making sure the build still works with the expected version, which it
won't if the version is updated upstream.

The version of the artifact needs to be changed due to Universal
Packages' constraints. It can only accept three sections, but with the
four from V8 and one from our builds, we'd have five.

Given that the first two sections in V8's version is the Chrome version
(ex. 9.4 -> 94), we join them. We also need to join the two last ones,
so that we can add a build ID.

The final versioning scheme is:
 * 9.4.146.17  ->  94.14617.$BuildId

This does create the problem where 14.617 and 146.17 yield the same
version on our side, but our builds of V8 will probably be few and far
between that we can worry about that when it stops us from progressing.

To download the latest 94.14617, use version "94.14617.*", or even the
latest 94, just use "94.*".