Fetch and improve V8 build
Adding a new script to fetch the pipeline artifact for using V8 into
CCF.

Mistakes on artifact naming have made it hard to get the right package.
When implementing the fetch script I realised the problem and have now
fixed in all three places: build-v8, pipeline, fetch-v8

Also, build now fetches the tag "$VERSION" instead of relying the
branch-head to be on the same version and failing.

A future addition would be to pass less info (ex. 9.4) and get the
script to fetch the head and figure out the version on its own.