VSCode C++ simplest configuration
Maps the configuration of the workspace to use some generic flags.
Other flags can be added on top, if needed for different files.

This helps the fact that compile_commands.json crashes the clangd
front-end in this project. The option, then, is to disable the
compilation database and have a generic command line.