Move key_exchange to src/crypto
Even though this uses tls' key exchange, this has nothing to do with our
TLS handling and should be in src/crypto folder instead.

Like other crypto logic, this depends on mbedTLS and needs to be
migrated to OpenSSL sometime soon, but not now. The main priority is to
get TLS 1.3 operational (with OpenSSL) and moving this out helps reduce
the workload for that job.