Silencing a warning for ignoring return value
ECDSA_SIG_set0 owns the pointer, so the `release` return value can be
ignored, but this is specific of this pattern and the Unique_BIGNUM
class itself should always return it and let the user decides.

This commit adds a comment (with reference) to make explicit why we're
ignoring the release return values.