Refactor Unique_OBJ classes in openssl_wrappers
When adding a few more unique pointers to that class, I realised the
patterns were so close to each other that we could just do with a
template wrapper and simplify the code repetition.

This change is a non-functional but should make it easier to add new
unique pointers here without repeating too much.