Move crypto tests in tls to crypto
Those two tests weren't testing anything related to TLS, but crypto
functions directly.

`tls_test` was a misnomer, just because Base64 used to be implemented
with MbedTLS only, so it was here. Moving to `base64_test` in `crypto`.

`cert_test` wasn't testing TLS certificates, but key pair certificates
that were implemented in `crypto` before we started the TLS cleanup.
Not sure what it was doing there, but moved to `kp_cert_test` in
`crypto`.

With this, the TLS `test` directory becomes empty, which is worrying.
In a following commit, I'll add some actual TLS tests, that will
eventually help me with testing OpenSSL's implementation without
resorting to full system tests (which right now are the only ones
catching bugs in any TLS implementation).