Simple TLS tests
We didn't have any tests checking the basic TLS functionality of the
src/tls implementation. This is an initial attempt at fixing that.

The idea is to have at least some basic checks to use when adding
OpenSSL functionality to implement TLS 1.3, so checks shouldn't depend
on mbedtls or openssl data structures.

There is only the handshake look that needs to know error values, but
this can be refactored once we bring OpenSSL into the picture.

This new test checks certificate creation, communication patterns,
including handshake, send and receive, close.