tlstest.sh runs multiple times, compares to golden
The tlstest.sh script fails if the tls_report.* files exist. This commit
removes the files, if they exist, before running the tests, to avoid
spurious failures during multiple re-runs of all tests on a local
machine.

Also adding golden file to tlstest to compare to output, filtering the
unstable output (IP address, hashes, comments, etc). On error, the test
prepares the current output, to be moved into the golden file if
correct.

This makes sure we actually check the output and make sure there aren't
any changes, or if there are, they need to be copied to the new golden
file and show in the diff of the commit.

Fixes #3365