Add more ciphers and restrict elliptic curves in OpenSSL
These are the approved ciphers and curves that we can support. The TLS
1.2 cipher remains the same for compatibility, but the TLS 1.3 cipher
suites is updated to include stronger ciphers.

Comparing the TLS Report, now that I know what to look for, this is now
on par with the previous report (same bit-rate, server-order, restricted
curves), but strictly better with TLS 1.3 and more, stronger approved
ciphers in 1.3.