Allow logging before enclave node creation to be seen
Before the enclave node is initialised, and the logger structure is
setup (sending logs through the ring buffer), any error ends up as
InternalError because we can't see the logs yet.

This commit changes that by creating the logging infrastructure before
the node is created (but after the ring buffer is setup), so that any
logs during that time, inside the enclave, can be flushed out and
displayed with the other error messages.

This allows us to be more verbose inside the enclave while constructing
the objects and creating the nodes, which helps debug the causes that
enclaves fail.

Extra work done to pass CreateNodeStatus around instead of throwing
exceptions, handling OE and CCF errors across enclave<->host barriers,
adding a new type of error (version mismatch) and adding a few traces
across the enclave constructor and node creations to help narrow issues
at a first glance.

Example: Creating a sandbox with the wrong parameter.

Command Line:
```
sandbox.sh -v \
  -p samples/apps/logging/liblogging \
  -n local://127.0.0.1:1234 \
  --subject_alt_names DNSName:foobar.azure.com
```

Note the alt-name is `DNSName` instead of `dNSName`. This will trigger
an error.

None of the output below was showing before this patch, on case of node
creation error, since the node still hadn't have time to setup the
writer and the host wasn't pulling any from the ring buffer.

Output:
```
[trace] ../src/enclave/enclave.h:108
  | Creating node
[trace] ../src/enclave/enclave.h:112
  | Creating context
[trace] ../src/enclave/enclave.h:121
  | Creating RPC actors / ffi
[trace] ./include/ccf/endpoint_registry.h:66
  | Parsed a templated endpoint: /log/private/raw_text/{id}
    became /log/private/raw_text/([^/]+)
[trace] ./include/ccf/endpoint_registry.h:69
  | Component names are: id
[trace] ../src/enclave/enclave.h:134
  | Initialize node
[debug] ../src/ds/state_machine.h:44
  | [NodeState] Advancing to state "Initialized" (from "Uninitialized")
[info ] ../src/enclave/rpc_sessions.h:164
  | Setting max open sessions on interface "primary_rpc_interface"
    (127.0.0.1:1234) to [1000, 1010]
[trace] ../src/enclave/enclave.h:177
  | Creating node with start_type Start
[fail ] ../src/enclave/enclave.h:182
  | Error starting node: SAN could not be parsed: DNSName:foobar.azure.com,
    must be (iPAddress|dNSName):VALUE
```

Fixes #857