Allow logging before enclave node creation to be seen
Before the enclave node is initialised, and the logger structure is
setup (sending logs through the ring buffer), any error ends up as
InternalError because we can't see the logs yet.

This commit changes that by creating the logging infrastructure before
the node is created (but after the ring buffer is setup), so that any
logs during that time, inside the enclave, can be flushed out and
displayed with the other error messages.

This allows us to be more verbose inside the enclave while constructing
the objects and creating the nodes, which helps debug the causes that
enclaves fail.

Extra work done to pass CreateNodeStatus around instead of throwing
exceptions, handling OE and CCF errors across enclave<->host barriers,
adding a new type of error (version mismatch) and adding a few traces
across the enclave constructor and node creations to help narrow issues
at a first glance.

Fixes #857