More TLS cleanups
When moving to OpenSSL, we left some options open for the day we
implement QUIC and need to reuse the TLS context.

However, while looking at QUIC implementation, we realised it doesn't
need some of the things we expected, for example, it doesn't support
DTLS.

There is an effort to port DTLS to QUIC [1], but that seems it's not
going anywhere, so we shouldn't wait for that to work, given this
is a simple cleanup. We can add later, of course.

[1] https://tools.ietf.org/id/draft-rescorla-quic-over-dtls-00.html

Also cleaning the debug BIO hook that is no longer used.