Allow sgx and virtual builds to be installed side-by-side using Ansible.
The Ansible playbook for SGX tries to install the open-enclave package,
whereas on Virtual builds it installs hostverify. Since these two are
mutually exclusive, they cannot be run on the same machine.

This changes the playbook to install hostverify only if OE is not
already installed. This allows the virtual build to be installed even
after the SGX one has been.

We also uninstall hostverify when trying to install OE, as this allows
the SGX build after the virtual one.

Unfortunately, it is still vital to install the SGX build first, then
the virtual one. When running the playbooks in the reverse order,
removing hostverify also removes ccf_virtual.