Allow applications to add .well-known endpoints.
The ACME frontend was binding to the entire `.well-known` directory,
intercepting any requests make to it, including to paths outside of the
particular `acme-challenge` sub-directory.

This was preventing applications from handling any paths inside the
`.well-known`, which may be necessary to implement certain protocols,
such as did:web.

The code to extract the actor name from an incoming request is special
cased for the .well-known directory, in which case it will include the
second path component in the actor's name. THe ACME frontend's actor is
now the full `.well-known/acme-challenge`.

We did consider implementing this by instead moving the ACME endpoints
into the CommonEndpointRegistry, indirectly making them part of the
application frontend, and removing the .well-known actor. Unfortunately
this would make the endpoints accessible only once the application
frontend has been opened, which would prevent a service from obtaining
an ACME certificate early in its lifecycle.