Add processing of host process stdio.
When an application launches a process on the host, it now has the
option to attach a bytestring which will be passed on the process'
standard input.

Additionally, the standard output and error of the process are captured
and printed in CCF's log. This makes it a lot easier to monitor and
diagnose issues in production.