Handle raft_driver quirk.
The "nodes" command in raft scenarios causes N consecutive "add_configuration" log lines to be emitted, where N is determined by the "nodes" parameter. At this stage, the nodes are in the "Pending" state. However, the enablement condition of "ccfraft!Timeout" is only true for nodes in the "Candidate" or "Follower" state. Therefore, we include this action to address this quirk in the raft_driver.