Remove gdb install
This is mainly prompted by the fact that apt update does not seem to work today as a result of the nodejs apt repo being in a bad state.

More generally though, imposing the time this takes on all users seems unwarranted. lldb works much better than gdb on our builds.