Run CodeQL in container
With the CI job now run in-container, and with executors being re-used in the private pool to improve efficiency, there end up being directories created by root (inside the container) that the checkout step on non-container jobs is unable to clean up.

One possible solution is to move all jobs run in private pools to container images. I remembered, probably incorrectly, that CodeQL could not be run in a container. But that does not seem to be the case now, there are only a few requirements: https://docs.github.com/en/code-security/code-scanning/creating-an-advanced-setup-for-code-scanning/running-codeql-code-scanning-in-a-container