Reduce max message size ring buffer in destruction test
The default was increased in #5866, which made the destruction test to
fail more ofter because of unexpected destruction flow. This was
intended to fail on SGX due to out-of-memory and on virtual due to
max ringbuffer message size overflow.

After increasing the message limit we get the primary to slow (in
debug build, at least), so election happens and is messing up the
network, and so test can fail on missing primary or other states
it doesn't expect.

We mitigate the issue here by rolling back msg limit, but the proper
fix might be to rewrite the test to make it more robust.